qcacld-2.0: Fix potential skb leakage
In commit Change-Id: Id880c5fe423eb0b2a2c01677d8fa7c4a784c74df
We add the case to exclude the support of the nl_srv for MULTI_IF_NAME.

The skb in the ucast and bcast transmit is not really sending to
kernel, so the driver itself has to free the skb.

Also move the MULTI_IF_NAME case to header file.

Change-Id: Iaa8bf34194050cef473b58ddc40f750d88eee26c
CRs-Fixed: 995209
Signed-off-by: Subhani Shaik <subhanis@codeaurora.org>