qcacld-2.0: Fix incorrect freeing of memory for internal oem response
Fix incorrect freeing of memory for internally generated oem data
response. For internal oem data response, response length will be
zero and oem_data_rsp will not be allocated any memory. Add a check
to not free that memory for internal data response and also assign
that pointer to NULL in the sender.

Change-Id: I2784b5f6f60988223baf82df025f98083575fd87
CRs-Fixed: 1004248
Signed-off-by: Subhani Shaik <subhanis@codeaurora.org>