bullhead_defconfig: disable CONFIG_PFT as it is unsupported
Disable the drivers/platform/msm/pft.c driver as it is unsupported by
Qualcomm, and opens us up to a wide range of potential attack surfaces
that has not been audited by anyone.

Qualcomm recommends that it be disabled, as it hooks into SELinux in
some "interesting" ways, and the userspace portion of the code is not
even part of the image.

Bug: 28588434
Change-Id: I932ff9f8e9bc3aff01585f210514f52958a508b4
Author: Greg Kroah-Hartman <gregkh@google.com>