qcald-2.0: prevent unloading driver when the new mode is same
We recently change the logic for loading and unload driver for mode
change (CL5999313).
However, with this code, in case of getting loading driver request
with same mode, we unload driver and return -INVAL.

for the scenario where we turn off tethering, we got loading request
for STA mode two times
At second request, we unload driver and return error.

Eventually, we failed to turn on WLAN automatically.

As a fix, we just return success in case of same mode request.

Bug: 28740503
Change-Id: I787db8eae0cf9c9a47a5dc06a3adbc2f31837711

Conflicts:
	drivers/staging/qcacld-2.0/CORE/HDD/src/wlan_hdd_main.c